fixing package name
