+++ package.json
@@ -60,7 +60,7 @@
     "transform-ast": "2.4.4",
     "typescript": "^5.7.3",
     "typescript-eslint": "^8.32.0",
-    "vite": "6.3.5",
+    "vite": "6.3.3",
     "vite-plugin-checker": "0.9.1",
     "workbox-build": "7.3.0",
     "workbox-core": "7.3.0",
@@ -119,7 +119,7 @@
       "strip-css-comments": "5.0.0",
       "transform-ast": "2.4.4",
       "typescript": "5.7.3",
-      "vite": "6.3.5",
+      "vite": "6.3.3",
       "vite-plugin-checker": "0.9.1",
       "workbox-build": "7.3.0",
       "workbox-core": "7.3.0",
