+++ pom.xml
@@ -4,7 +4,7 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.example.application</groupId>
+    <groupId>com.bobbysoft.application</groupId>
     <artifactId>bobbysoft-organizer</artifactId>
     <version>1.0-SNAPSHOT</version>
