+++ src/main/java/com/bobbysoft/application/base/domain/package-info.java
@@ -1,0 +1,9 @@
+/**
+ * This package contains reusable domain classes.
+ */
+@NullMarked
+@NonNullApi // Until https://github.com/vaadin/hilla/issues/2612 has been fixed
+package com.bobbysoft.application.base.domain;
+
+import org.jspecify.annotations.NullMarked;
+import org.springframework.lang.NonNullApi;
