+++ src/main/java/com/bobbysoft/application/taskmanagement/service/package-info.java
@@ -1,0 +1,13 @@
+/**
+ * This package contains the application services of the Task Management sample feature.
+ * <p>
+ * You can add as many application services as you want to this package, as long as they belong to the same feature.
+ * If any of them use Data Transfer Objects (DTO), you can add them here as well.
+ * </p>
+ */
+@NullMarked
+@NonNullApi // Until https://github.com/vaadin/hilla/issues/2612 has been fixed
+package com.bobbysoft.application.taskmanagement.service;
+
+import org.jspecify.annotations.NullMarked;
+import org.springframework.lang.NonNullApi;
