+++ src/main/java/com/bobbysoft/application/usermanagement/domain/UserRepository.java
@@ -1,0 +1,9 @@
+package com.bobbysoft.application.usermanagement.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
+
+import java.util.UUID;
+
+public interface UserRepository extends JpaRepository<UserEntity, UUID>, JpaSpecificationExecutor<UserEntity> {
+}
