+++ src/main/java/com/bobbysoft/application/usermanagement/dto/UserInfoDto.java
@@ -1,0 +1,6 @@
+package com.bobbysoft.application.usermanagement.dto;
+
+import java.util.Collection;
+
+public record UserInfoDto(String name, Collection<String> authorities, String email, String avatarUrl) {
+}
