--- src/main/java/com/example/application/base/domain/package-info.java
@@ -1,9 +1,0 @@
-/**
- * This package contains reusable domain classes.
- */
-@NullMarked
-@NonNullApi // Until https://github.com/vaadin/hilla/issues/2612 has been fixed
-package com.example.application.base.domain;
-
-import org.jspecify.annotations.NullMarked;
-import org.springframework.lang.NonNullApi;
