--- src/main/java/com/example/application/taskmanagement/domain/Task.java
@@ -1,63 +1,0 @@
-package com.example.application.taskmanagement.domain;
-
-import com.example.application.base.domain.AbstractEntity;
-import com.fasterxml.jackson.annotation.JsonProperty;
-import jakarta.persistence.*;
-import jakarta.validation.constraints.Size;
-import org.jspecify.annotations.Nullable;
-
-import java.time.Instant;
-import java.time.LocalDate;
-
-@Entity
-@Table(name = "task")
-public class Task extends AbstractEntity<Long> {
-
-    public static final int DESCRIPTION_MAX_LENGTH = 255;
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "task_id")
-    @JsonProperty
-    private Long id;
-
-    @Column(name = "description", nullable = false, length = DESCRIPTION_MAX_LENGTH)
-    @Size(max = DESCRIPTION_MAX_LENGTH)
-    private String description;
-
-    @Column(name = "creation_date", nullable = false)
-    private Instant creationDate;
-
-    @Column(name = "due_date")
-    @Nullable
-    private LocalDate dueDate;
-
-    @Override
-    public @Nullable Long getId() {
-        return id;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public Instant getCreationDate() {
-        return creationDate;
-    }
-
-    public void setCreationDate(Instant creationDate) {
-        this.creationDate = creationDate;
-    }
-
-    public @Nullable LocalDate getDueDate() {
-        return dueDate;
-    }
-
-    public void setDueDate(@Nullable LocalDate dueDate) {
-        this.dueDate = dueDate;
-    }
-}
