--- src/main/java/com/example/application/taskmanagement/domain/package-info.java
@@ -1,16 +1,0 @@
-/**
- * This package contains the domain model of the Task Management sample feature.
- * <p>
- * You can add as many domain model artifacts (such as entities, value objects, repositories, domain events,
- * and domain services) to this package, as long as they belong to the same feature.
- * </p>
- * <p>
- * If you have domain classes that are re-usable across multiple features, add them to the {@code base.domain} package.
- * </p>
- */
-@NullMarked
-@NonNullApi // Until https://github.com/vaadin/hilla/issues/2612 has been fixed
-package com.example.application.taskmanagement.domain;
-
-import org.jspecify.annotations.NullMarked;
-import org.springframework.lang.NonNullApi;
