--- src/main/java/com/example/application/taskmanagement/service/TaskService.java
@@ -1,45 +1,0 @@
-package com.example.application.taskmanagement.service;
-
-import com.example.application.taskmanagement.domain.Task;
-import com.example.application.taskmanagement.domain.TaskRepository;
-import com.vaadin.flow.server.auth.AnonymousAllowed;
-import com.vaadin.hilla.BrowserCallable;
-import org.jspecify.annotations.Nullable;
-import org.springframework.data.domain.Pageable;
-import org.springframework.transaction.annotation.Propagation;
-import org.springframework.transaction.annotation.Transactional;
-
-import java.time.Clock;
-import java.time.LocalDate;
-import java.util.List;
-
-@BrowserCallable
-@AnonymousAllowed
-@Transactional(propagation = Propagation.REQUIRES_NEW)
-public class TaskService {
-
-    private final TaskRepository taskRepository;
-
-    private final Clock clock;
-
-    TaskService(TaskRepository taskRepository, Clock clock) {
-        this.taskRepository = taskRepository;
-        this.clock = clock;
-    }
-
-    public void createTask(String description, @Nullable LocalDate dueDate) {
-        if ("fail".equals(description)) {
-            throw new RuntimeException("This is for testing the error handler");
-        }
-        var task = new Task();
-        task.setDescription(description);
-        task.setCreationDate(clock.instant());
-        task.setDueDate(dueDate);
-        taskRepository.saveAndFlush(task);
-    }
-
-    public List<Task> list(Pageable pageable) {
-        return taskRepository.findAllBy(pageable).toList();
-    }
-
-}
