--- src/main/java/com/example/application/usermanagement/domain/UserRepository.java
@@ -1,9 +1,0 @@
-package com.example.application.usermanagement.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
-
-import java.util.UUID;
-
-public interface UserRepository extends JpaRepository<UserEntity, UUID>, JpaSpecificationExecutor<UserEntity> {
-}
