--- src/main/java/com/example/application/usermanagement/dto/UserInfoDto.java
@@ -1,6 +1,0 @@
-package com.example.application.usermanagement.dto;
-
-import java.util.Collection;
-
-public record UserInfoDto(String name, Collection<String> authorities, String email, String avatarUrl) {
-}
