--- src/main/java/com/example/application/usermanagement/dto/validator/PasswordConfirmationValidator.java
@@ -1,20 +1,0 @@
-package com.example.application.usermanagement.dto.validator;
-
-import com.example.application.usermanagement.dto.UserRegistrationDto;
-import jakarta.validation.ConstraintValidator;
-import jakarta.validation.ConstraintValidatorContext;
-
-public class PasswordConfirmationValidator implements ConstraintValidator<ValidPasswordConfirmation, UserRegistrationDto> {
-    @Override
-    public void initialize(ValidPasswordConfirmation constraintAnnotation) {
-        ConstraintValidator.super.initialize(constraintAnnotation);
-    }
-
-    @Override
-    public boolean isValid(UserRegistrationDto userRegistrationDto, ConstraintValidatorContext constraintValidatorContext) {
-        String password = userRegistrationDto.password();
-        String confirmation = userRegistrationDto.passwordConfirmation();
-
-        return password.equals(confirmation);
-    }
-}
