+++ src/main/resources/application.properties
@@ -1,14 +1,11 @@
 server.port=${PORT:8080}
 logging.level.org.atmosphere=warn
-
 # Launch the default browser when starting the application in development mode
 vaadin.launch-browser=true
 # To improve the performance during development.
 # For more information https://vaadin.com/docs/latest/flow/integrations/spring/configuration#special-configuration-parameters
-vaadin.allowed-packages=com.vaadin,org.vaadin,com.example.application
-
+vaadin.allowed-packages=com.vaadin,org.vaadin,com.bobbysoft.application
 # Open-in-view is only needed if you use lazy-loaded entities in your Flow views.
 spring.jpa.open-in-view=false
-
 # Initialize the JPA Entity Manager before considering data.sql so that the EM can create the schema and data.sql contain data
-spring.jpa.defer-datasource-initialization = true
+spring.jpa.defer-datasource-initialization=true
